+++ CHANGELOG.md
@@ -6,14 +6,18 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- .rdm animation support
 - `<Materials>` and `<MaterialLODInfos>` import
 - issue with yellow metal texture from Blender export
 - support and strip comments from `.cfg`, `.ifo`
 - replace Cesium glb/gltf conversion with much simpler https://github.com/najadojo/gltf-import-export
 - Right-click `Import Materials`. Reorders material list and fills some gaps.
 - support glTF binary in `gltf` converter
+- Option to move files to `maps`, `rdm` and `anim`
+
+### Added
+
 - Show model animation in .cfg outline
+- Support animations in glTF conversion.
 
 ## [0.1.1]